### 2020-07-23 #2509 @whedon
[PRE REVIEW]: MLJ: A Julia package for composable machine learning
**Submitting author:** @ablaom (<a href="http://orcid.org/0000-0001-6689-886X">Anthony Blaom</a>)
**Repository:** <a href="https://github.com/alan-turing-institute/MLJ.jl" target ="_blank">https://github.com/alan-turing-institute/MLJ.jl</a>
**Version:** v0.11.6
