### 2020-07-23 #2509 @Kevin-Mattheus-Moerman
@ablaom thanks for your submission. I will be looking for editors to handle this work. In the mean time I've quickly looked at your paper. It seems rather longer than our typical paper. Please review our [criteria](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) and consider shortening it where possible. Our system also didn't detect a "Statement of need", you may have text covering similar content but consider collecting it concisely under this heading.
