### 2020-07-23 f917fa2
rm deprecated kind=... syntax from paper.md
