### 2020-07-24 #2509 @ablaom
Dear Yuan Tang,

Thanks for that. The following two people on your list include machine learning as domains of expertise. And I am not personally connected to either.

vchuravy

ahwililia

Cheers,
Anthony
