### 2020-07-24 #2509 @ahwillia
Thanks, I'd be happy to provide a review for this package. I talked with @degleris1 and he has agreed to help me write my review -- he's a current grad student at Stanford with lots of experience with Julia and ML (we've collaborated in the past). This package is quite large and I know I could use his help. Also you should consider adding him to your list of reviewers for future Julia submissions, and this would be good experience for him to see the process.
