### 2020-07-24 733bf8a
switch to new @from_network synthax in docs
