### 2020-08-03 31332df
add links to JuliaCon 2020 workshop
