### 2020-08-06 d711574
add Generating Synthetic Data section to manual
