### 2020-08-07 5500e20
fix readme links for manual
