### 2020-08-09 27a042b
Document the `JointProbabilistic` supervised model type
