### 2020-08-12 069857e
corrected wrong method signatures in docs
