### 2020-08-13 a4a3e10
Discourage fields with type Union{Nothing,T}
Discourage fields with type Union{Nothing,T} when adding models
