### 2020-08-13 40e00a0
add to quick state guide
