### 2020-08-13 282b4bb
add comment to adding new models section
