### 2020-08-14 36fd38c
readme fixes from #634
