### 2020-08-14 ab214d7
replace link to wow.ipynb to non-binder version
