### 2020-08-18 13a0f24
Make sure that CompatHelper opens pull requests against the `dev` branch
