### 2020-08-25 27674bf
add Sneak Preview section to manual Introduction
