+++ docs/src/adding_models_for_general_use.md
@@ -148,7 +148,7 @@ end
 ```
 
 *Important.* The clean method must have the property that
-`clean!(clean!(model)) == clean!(model)` for any instance `model`. 
+`clean!(clean!(model)) == clean!(model)` for any instance `model`.
 
 Although not essential, try to avoid `Union` types for model
 fields. For example, a field declaration `features::Vector{Symbol}`
@@ -505,7 +505,7 @@ This object automatically assigns zero-probability to the unseen class
 `:rare` (i.e., `pdf.(yhat, :rare)` works and returns a zero
 vector). If you would like to assign `:rare` non-zero probabilities,
 simply add it to the first vector (the *support*) and supply a larger
-`probs` matrix. 
+`probs` matrix.
 
 If instead of raw labels `[:a, :b]` you have the corresponding
 `CategoricalElement`s (from, e.g., `filter(cv->cv in unique(y),
@@ -653,7 +653,7 @@ For example, a three parameter model of the form
 ```julia
 mutable struct MyModel{D} <: Deterministic
     alpha::Float64
-	beta::Int
+        beta::Int
     distribution::D
 end
 ```
@@ -662,8 +662,8 @@ you might declare (order matters):
 ```julia
 MMI.hyperparameter_ranges(::Type{<:MyModel}) =
     (range(Float64, :alpha, lower=0, upper=1, scale=:log),
-	 range(Int, :beta, lower=1, upper=Inf, origin=100, unit=50, scale=:log),
-	 nothing)
+         range(Int, :beta, lower=1, upper=Inf, origin=100, unit=50, scale=:log),
+         nothing)
 ```
 
 Here is the complete list of trait function declarations for `DecisionTreeClassifier`
@@ -685,7 +685,7 @@ Alternatively these traits can also be declared using `MMI.metadata_pkg` and `MM
 MMI.metadata_pkg(DecisionTreeClassifier,name="DecisionTree",
                      uuid="7806a523-6efd-50cb-b5f6-3fa6f1930dbb",
                      url="https://github.com/bensadeghi/DecisionTree.jl",
-                     julia=true)   
+                     julia=true)
 
 MMI.metadata_model(DecisionTreeClassifier,
                         input=MMI.Table(MMI.Continuous),
@@ -693,7 +693,7 @@ MMI.metadata_model(DecisionTreeClassifier,
                         path="MLJModels.DecisionTree_.DecisionTreeClassifier")
 ```
 
-*Important.* Do not omit the `path` specifcation. 
+*Important.* Do not omit the `path` specifcation.
 
 ```@docs
 MMI.metadata_pkg
@@ -759,8 +759,8 @@ implementation should define a value for the `output_scitype` trait. A
 declaration
 
 ```julia
-output_scitype(::Type{<:SomeSupervisedModel}) = T 
-``` 
+output_scitype(::Type{<:SomeSupervisedModel}) = T
+```
 
 is an assurance that `scitype(transform(model, fitresult, Xnew)) <: T`
 always holds, for any `model` of type `SomeSupervisedModel`.
@@ -783,82 +783,93 @@ similar fashion. The main differences are:
   the same return value `(fitresult, cache, report)`. An `update`
   method (e.g., for iterative models) can be optionally implemented in
   the same way.
-  
+
 - A `transform` method is compulsory and has the same signature as
-  `predict`, as in `MLJModelInterface.transform(model, fitresult, Xnew)`. 
-  
+  `predict`, as in `MLJModelInterface.transform(model, fitresult, Xnew)`.
+
 - Instead of defining the `target_scitype` trait, one declares an
   `output_scitype` trait (see above for the meaning).
 
[…]
