### 2020-08-25 9c9758c
add Sneak Preview section to manual Introduction
