### 2020-08-25 33bb227
Revamp manual entry for distribution-fitting models #641
fix bugs in revamp
