### 2020-08-25 e279e51
Revamp manual entry for distribution-fitting models #641
fix bugs in revamp
