### 2020-08-27 9f60af3
Add docs for JointProbabilistic
small fixes
