+++ docs/src/adding_models_for_general_use.md
@@ -183,13 +183,30 @@ parameters.
 
 ## Supervised models
 
+### Mathematical assumptions
+
+At present, MLJ's performance estimate functionality (resampling using
+`evaluate`/`evaluate!`) tacitly assumes that feature-label pairs of
+observations `(X1, y1), (X2, y2), (X2, y2), ...` are being modelled as
+identically independent random variables (i.i.d.), and constructs some
+kind of representation of an estimate of the conditional probablility
+`p(y | X)` (`y` and `X` single observations). It may be that a model
+implementing the MLJ interface has the potential to make predictions
+under weaker assumptions (e.g., time series forecasting
+models). However the output of the compulsory `predict` method
+described below should be the output of the model under the i.i.d
+assumption.
+
+In the future newer methods may be introduced to handle weaker
+assumptions (see, e.g., [Predicting joint distributions](@ref) below).
+
+
+### Summary of methods
+
 The compulsory and optional methods to be implemented for each
 concrete type `SomeSupervisedModel <: MMI.Supervised` are
 summarized below. An `=` indicates the return value for a fallback
 version of the method.
-
-
-### Summary of methods
 
 Compulsory:
 
@@ -393,17 +410,27 @@ The fallback is to return `(fitresult=fitresult,)`.
 ### The predict method
 
 A compulsory `predict` method has the form
+
 ```julia
 MMI.predict(model::SomeSupervisedModel, fitresult, Xnew) -> yhat
 ```
 
-Here `Xnew` will have the same form as the `X` passed to `fit`.
+Here `Xnew` will have the same form as the `X` passed to
+`fit`. 
+
+Note that while `Xnew` generally consists of multiple observations
+(e.g., has multiple rows in the case of a table) it is assumed, in view of
+the i.i.d assumption recalled above, that calling `predict(..., Xnew)`
+is equivalent to broadcasting some method `predict_one(..., x)` over
+the individual observations `x` in `Xnew` (a method implementing the
+probablility distribution `p(X |y)` above).
+
 
 #### Prediction types for deterministic responses.
 
 In the case of `Deterministic` models, `yhat` should have the same
 scitype as the `y` passed to `fit` (see above). Any `CategoricalValue`
-or `CategoricalString` elements of `yhat` **must have a pool == to the
+elements of `yhat` **must have a pool == to the
 pool of the target `y` presented in training**, even if not all levels
 appear in the training data or prediction itself. For example, in the
 case of a univariate target, such as `scitype(y) <:
@@ -412,14 +439,13 @@ MLJ.classes(y[j])` for all admissible `i` and `j`. (The method
 `classes` is described under [Convenience methods](@ref) below).
 
 Unfortunately, code not written with the preservation of categorical
-levels in mind poses special problems. To help with this, MLJModelInterface
-provides three utility methods: `int` (for converting a
-`CategoricalValue` or `CategoricalString` into an integer, the
-ordering of these integers being consistent with that of the pool),
-`decoder` (for constructing a callable object that decodes the
-integers back into `CategoricalValue`/`CategoricalString` objects),
-and `classes`, for extracting all the `CategoricalValue` or
-`CategoricalString` objects sharing the pool of a particular
+levels in mind poses special problems. To help with this,
+MLJModelInterface provides three utility methods: `int` (for
+converting a `CategoricalValue` into an integer, the ordering of these
[…]
