### 2020-08-28 b7c5ea4
Update documentation around `return!` syntax
