### 2020-08-28 793f90b
add example to Method II composite model export
