+++ .travis.yml
@@ -6,7 +6,8 @@ julia:
   - 1.0
   - 1.1
   - 1.2
-  - 1.3
+  #- 1.3
+  - 1.5
 notifications:
   email: false
 git:
