### 2020-08-28 d66dead
rm 1.1 and 1.2 from ci
