+++ .travis.yml
@@ -4,8 +4,8 @@ os:
   - linux
 julia:
   - 1.0
-  - 1.1
-  - 1.2
+  #- 1.1
+  #- 1.2
   #- 1.3
   - 1.5
 notifications:
