+++ Project.toml
@@ -25,7 +25,7 @@ CategoricalArrays = "^0.8"
 ComputationalResources = "^0.3"
 Distributions = "^0.21,^0.22,^0.23"
 MLJBase = "^0.14.9"
-MLJModels = "^0.11"
+MLJModels = "^0.11, 0.12"
 MLJScientificTypes = "^0.2.1"
 MLJTuning = "^0.4"
 ProgressMeter = "^1.1"
