### 2020-08-29 8fa0180
CompatHelper: bump compat for "MLJBase" to "0.15"
