+++ Project.toml
@@ -24,7 +24,7 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 CategoricalArrays = "^0.8"
 ComputationalResources = "^0.3"
 Distributions = "^0.21,^0.22,^0.23"
-MLJBase = "^0.14.9"
+MLJBase = "^0.14.9, 0.15"
 MLJModels = "^0.11"
 MLJScientificTypes = "^0.2.1"
 MLJTuning = "^0.4"
