### 2020-08-29 7764ce4
CompatHelper: bump compat for "MLJScientificTypes" to "0.3"
