+++ Project.toml
@@ -26,7 +26,7 @@ ComputationalResources = "^0.3"
 Distributions = "^0.21,^0.22,^0.23"
 MLJBase = "^0.14.9"
 MLJModels = "^0.11"
-MLJScientificTypes = "^0.2.1"
+MLJScientificTypes = "^0.2.1, 0.3"
 MLJTuning = "^0.4"
 ProgressMeter = "^1.1"
 StatsBase = "^0.32,^0.33"
