### 2020-08-31 7a99a44
Add some more fixes addressing #379 and for machine constructor
more

more
