### 2020-09-01 d6db5ee
more minor doc-fixes
