### 2020-09-01 a93ed62
fix remaining warnings in doc generation
