+++ Project.toml
@@ -27,7 +27,7 @@ Distributions = "^0.21,^0.22,^0.23"
 MLJBase = "^0.15.1"
 MLJModels = "^0.12.1"
 MLJScientificTypes = "^0.3.0"
-MLJTuning = "^0.4"
+MLJTuning = "^0.5"
 ProgressMeter = "^1.1"
 StatsBase = "^0.32,^0.33"
 Tables = "^0.2,^1.0"
