+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.14",
   "v0.13",
   "v0.12",
   "v0.11",
