### 2020-09-28 #2704 @whedon
[REVIEW]: MLJ: A Julia package for composable machine learning
**Submitting author:** @ablaom (<a href="http://orcid.org/0000-0001-6689-886X">Anthony Blaom</a>)
**Repository:** <a href="https://github.com/alan-turing-institute/MLJ.jl" target ="_blank">https://github.com/alan-turing-institute/MLJ.jl</a>
**Version:** v0.14.1
