### 2020-10-01 dbdf397
address #654 document fix
