+++ docs/src/composing_models.md
@@ -519,6 +519,9 @@ Notes:
   calls `fit!` on the learning network machine `mach` and splits it
   into various pieces, as required by the MLJ model interface. See
   also the [`return!`](@ref) doc-string.
+  
+- **Important note** An MLJ `fit` method is not allowed to mutate its
+  `model` argument. 
 
 > **What's going on here?** MLJ's machine interface is built atop a more primitive *[model](simple_user_defined_models.md)* interface, implemented for each algorithm. Each supervised model type (eg, `RidgeRegressor`) requires model `fit` and `predict` methods, which are called by the corresponding *machine* `fit!` and `predict` methods. We don't need to define a  model `predict` method here because MLJ provides a fallback which simply calls the `predict` on the learning network machine created in the `fit` method. 
 
