### 2020-10-01 d9ec549
Fix missing fit_only! ref in docs
