+++ docs/src/composing_models.md
@@ -753,5 +753,5 @@ node
 return!
 ```
 
-See more on fitting nodes at [`fit!`](@ref) and [`fit_only!](@ref). 
+See more on fitting nodes at [`fit!`](@ref) and [`fit_only!`](@ref). 
 
