### 2020-10-01 9e883a1
address #663 (doc update)
