### 2020-10-01 c0ab77f
address #649 doc update
