### 2020-10-02 #2704 @henrykironde
**Contribution and authorship:**

`What cliteria was used for authorship selection?`

**License:**

`Could you edit the licence file and remove  '>'`

**The paper should be between 250-1000 words:**

`I will need to obain a recommendation for the paper length from the team.`

**Installation instructions:**
```
These can be made more clear.
The Installation instructions shown in the readme file after
the intro tend to usually clear to users.
```

**List of dependencies?**
```
The only issue with the dependencies has been stated in the README,
this makes users know of the status when installing 👍
```

**Example usage:**

```
I expect a user to be able to run a sample example by reading the `README` file in a few minutes.
I recommend that you move the `List of Wrapped Models` to any other location in the docs, but provide the link on the `README`
Then focus on a sample run/s. I have tried to run the code and faced some difficulties:

```

**Functionality documentation:**

```
Functions should have docs strings which are automatically turned into documentation.
A good example is listed
```
https://github.com/JuliaLang/Statistics.jl/blob/b384104d35ff0e7cf311485607b177223ed72b9a/src/Statistics.jl#L19

**Community guidelines:**

```
I should be able to see a developer doc section.
In case I want to clone this repo and change a few things.
I should be able to install the source from the current directory and there after running the code with the changes.
I expect the user to also have guidelines on how to run tests.
```

I have provided a sample run here hope we can make it smooth,

1) Package fails to install from GitHub

```j
docker run -it --rm julia
Unable to find image 'julia:latest' locally
latest: Pulling from library/julia
d121f8d1c412: Pull complete
a27e46e0dbc6: Pull complete
b6a50acaca53: Pull complete
Digest: sha256:f8a867690b5e341a66b20af1b0c09216367805e0eb87f6e059c13b683804b681
Status: Downloaded newer image for julia:latest
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 1.5.2 (2020-09-23)
 _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org/ release
|__/                   |

julia> using Pkg;

julia> Pkg.develop(PackageSpec(url="https://github.com/tlienart/OpenSpecFun_jll.jl"))
    Cloning git-repo `https://github.com/tlienart/OpenSpecFun_jll.jl`
  Resolving package versions...
 Installing known registries into `~/.julia`
######################################################################## 100.0%
      Added registry `General` to `~/.julia/registries/General`
Downloading artifact: OpenSpecFun
Updating `~/.julia/environments/v1.5/Project.toml`
  [efe28fd5] + OpenSpecFun_jll v0.5.3+1 `~/.julia/dev/OpenSpecFun_jll`
Updating `~/.julia/environments/v1.5/Manifest.toml`
  [efe28fd5] + OpenSpecFun_jll v0.5.3+1 `~/.julia/dev/OpenSpecFun_jll`
  [2a0f44e3] + Base64
  [ade2ca70] + Dates
 	.....
  [4ec0a83e] + Unicode

julia> using MLJ
ERROR: ArgumentError: Package MLJ not found in current path:
- Run `import Pkg; Pkg.add("MLJ")` to install the MLJ package.

Stacktrace:
 [1] require(::Module, ::Symbol) at ./loading.jl:893
```

2) Package installed from the release but example does not mention `Pkg.add("EvoTrees")`

```j
julia> import Pkg; Pkg.add("MLJ")
   Updating registry at `~/.julia/registries/General`
  Resolving package versions...
  Installed MLJModelInterface ──────────── v0.3.5
  .....
julia> using MLJ
[ Info: Precompiling MLJ [add582a8-e3ab-11e8-2d5e-e98b27df1bc7]
[ Info: Model metadata loaded from registry.

julia> X, y = @load_reduced_ames;

julia> booster = @load EvoTreeRegressor

ERROR: LoadError: ArgumentError: Package EvoTrees not found in current path:
- Run `import Pkg; Pkg.add("EvoTrees")` to install the EvoTrees package.
```

After a while, I managed to set up the example correct

```
julia> using MLJ
[ Info: Precompiling MLJ [add582a8-e3ab-11e8-2d5e-e98b27df1bc7]
[ Info: Model metadata loaded from registry.

julia> Pkg.add("EvoTrees")
  Resolving package versions...
  Installed CEnum ────────── v0.4.1
  ...........
  Installed GPUArrays ────── v5.2.1
  Installed CUDA ─────────── v1.3.3
Updating `~/.julia/environments/v1.5/Project.toml`
  [f6006082] + EvoTrees v0.5.0
Updating `~/.julia/environments/v1.5/Manifest.toml`
  [621f4979] + AbstractFFTs v0.5.0
  ........
  [872c559c] + NNlib v0.7.4
  [a759f4b9] + TimerOutputs v0.5.6

julia> X, y = @load_reduced_ames;

julia> booster = @load EvoTreeRegressor
[ Info: Precompiling EvoTrees [f6006082-12f8-11e9-0c9c-0d5d367ab1e5]
EvoTreeRegressor(
    loss = EvoTrees.Linear(),
    nrounds = 10,
    λ = 0.0f0,
    γ = 0.0f0,
    η = 0.1f0,
    max_depth = 5,
    min_weight = 1.0f0,
    rowsample = 1.0f0,
    colsample = 1.0f0,
    nbins = 64,
    α = 0.5f0,
    metric = :mse,
    rng = MersenneTwister(UInt32[0x000001bc]) @ 1002) @281

julia> booster.max_depth = 2
2

julia> booster.nrounds=50
50

julia> pipe = @pipeline ContinuousEncoder booster
Pipeline255(
    continuous_encoder = ContinuousEncoder(
            drop_last = false,
            one_hot_ordered_factors = false),
    evo_tree_regressor = EvoTreeRegressor(
            loss = EvoTrees.Linear(),
            nrounds = 50,
            λ = 0.0f0,
            γ = 0.0f0,
            η = 0.1f0,
            max_depth = 2,
            min_weight = 1.0f0,
            rowsample = 1.0f0,
            colsample = 1.0f0,
            nbins = 64,
            α = 0.5f0,
            metric = :mse,
            rng = MersenneTwister(UInt32[0x000001bc]) @ 1002)) @701

julia> max_depth_range = range(pipe,
                               :(evo_tree_regressor.max_depth),
                               lower = 1,
                               upper = 10)
MLJBase.NumericRange(Int64, :(evo_tree_regressor.max_depth), ... )

julia> self_tuning_pipe = TunedModel(model=pipe,
                                     tuning=RandomSearch(),
                                     ranges = max_depth_range,
                                     resampling=CV(nfolds=3, rng=456),
                                     measure=l1,
                                     acceleration=CPUThreads(),
                                     n=50)
DeterministicTunedModel(
    model = Pipeline255(
            continuous_encoder = ContinuousEncoder @343,
            evo_tree_regressor = EvoTreeRegressor{Float32,…} @281),
    tuning = RandomSearch(
            bounded = Distributions.Uniform,
            positive_unbounded = Distributions.Gamma,
            other = Distributions.Normal,
            rng = Random._GLOBAL_RNG()),
    resampling = CV(
            nfolds = 3,
            shuffle = true,
            rng = MersenneTwister(UInt32[0x000001c8]) @ 1002),
    measure = l1(),
    weights = nothing,
    operation = MLJModelInterface.predict,
    range = NumericRange(
            field = :(evo_tree_regressor.max_depth),
            lower = 1,
            upper = 10,
            origin = 5.5,
            unit = 4.5,
            scale = :linear),
    selection_heuristic = MLJTuning.NaiveSelection(nothing),
    train_best = true,
    repeats = 1,
    n = 50,
    acceleration = CPUThreads{Int64}(1),
    acceleration_resampling = CPU1{Nothing}(nothing),
    check_measure = true) @976

julia> mach = machine(self_tuning_pipe, X, y)
Machine{DeterministicTunedModel{RandomSearch,…}} @419 trained 0 times.
  args:
    1:	Source @746 ⏎ `Table{Union{AbstractArray{Continuous,1}, AbstractArray{Count,1}, AbstractArray{Multiclass{25},1}, AbstractArray{Multiclass{15},1}, AbstractArray{OrderedFactor{10},1}}}`
    2:	Source @854 ⏎ `AbstractArray{Continuous,1}`

julia> evaluate!(mach,
                       measures=[l1, l2],
                       resampling=CV(nfolds=6, rng=123),
                       acceleration=CPUProcesses(), verbosity=2)
[ Info: Distributing evaluations among 1 workers.

[ Info: Training Machine{DeterministicTunedModel{RandomSearch,…}} @419.
 .........
Evaluating over 50 metamodels: 100%[=========================] Time: 0:00:11
Evaluating over 6 folds: 100%[=========================] Time: 0:02:26
┌───────────┬───────────────┬────────────────────────────────────────────────────────┐
│ _.measure │ _.measurement │ _.per_fold                                             │
├───────────┼───────────────┼────────────────────────────────────────────────────────┤
│ l1        │ 16700.0       │ [17000.0, 16400.0, 14500.0, 16200.0, 16400.0, 19500.0] │
│ l2        │ 6.35e8        │ [6.13e8, 6.81e8, 4.35e8, 5.63e8, 5.98e8, 9.18e8]       │
└───────────┴───────────────┴────────────────────────────────────────────────────────┘
_.per_observation = [[[27700.0, 21400.0, ..., 11200.0], [12100.0, 1330.0, ..., 13200.0], [6490.0, 22000.0, ..., 13800.0], [12400.0, 7140.0, ..., 13000.0], [50800.0, 22700.0, ..., 1550.0], [32800.0, 4940.0, ..., 1110.0]], [[7.68e8, 4.59e8, ..., 1.26e8], [1.46e8, 1.77e6, ..., 1.73e8], [4.22e7, 4.86e8, ..., 1.9e8], [1.55e8, 5.09e7, ..., 1.7e8], [2.58e9, 5.13e8, ..., 2.42e6], [1.07e9, 2.44e7, ..., 1.24e6]]]
_.fitted_params_per_fold = [ … ]
_.report_per_fold = [ … ]
julia>
```

If we set up the docs well, we can reduce the time users spend when they land on the project to getting sample runs
