### 2020-10-09 #2704 @degleris1
Just finished my review. The package looks fantastic. I love the idea of scientific types. Everything (installation, tests, simple demos, defining a custom model) worked smoothly on my machine.

The paper does seem a bit long for JOSS. Could you say why?

Some really minor comments. These are things that you by no means have to address.

- As has been pointed out, the installation instructions are 2 link clicks away from the README. I don't have a problem with it, but it might be nice to have something right in the README, e.g.
> Open Julia and press `]` to open the package manager, then run
> `pkg> add MLJ`
> to install MLJ.

- In the [FAQ](https://alan-turing-institute.github.io/MLJ.jl/dev/frequently_asked_questions/), I noticed you used "his code" to refer to the user. It might be nice to change it to "their code" or "the user's code." As far as I can tell, it's actually the only gendered pronoun in entire docs / paper.

- Is there a good reason why I have to type `verbosity` to `Int`? It seems a little bit random from a user perspective, but I'm sure there's a good reason for it.
