### 2020-10-11 55dce29
edit LICENSE (formatting https://github.com/openjournals/joss-reviews/issues/2704#issuecomment-702969999
