### 2020-10-12 41ee9b2
Include self-contained installation instructions in lightning tour
