### 2020-10-12 1f3ef6b
in docs: drop requirment to annotate the type of `verbosity`
