### 2020-10-12 49568f2
add to docs: how to test MLJ installation
