### 2020-10-12 ce614cc
Move list of models to gh pages manual; refocus README.md
