+++ README.md
@@ -35,6 +35,14 @@
 
 **New to MLJ? Start [here](https://alan-turing-institute.github.io/MLJ.jl/dev/)**. 
 
+**Wanting to integrate an existing machine learning model into the MLJ
+framework?** Start
+[here](https://alan-turing-institute.github.io/MLJ.jl/dev/quick_start_guide_to_adding_models/).
+
+The remainining information on this page will be of interest primarily
+to developers interested in constributing to core packages in the MLJ
+ecosystem, whose organization is described further below.
+
 MLJ (Machine Learning in Julia) is a toolbox written in Julia
 providing a common interface and meta-algorithms for selecting,
 tuning, evaluating, composing and comparing machine learning models written in Julia and other languages.  MLJ is released
@@ -44,8 +52,8 @@ Institute](https://www.turing.ac.uk/).
 <br>
 <p align="center">
 <a href="#the-mlj-universe">MLJ Universe</a> &nbsp;•&nbsp; 
-<a href="#list-of-wrapped-models">List of Wrapped Models</a> &nbsp;•&nbsp;
 <a href="#known-issues">Known Issues</a> &nbsp;•&nbsp;
+<a href="#customizing-behavior">Customizing Behavior</a> &nbsp;•&nbsp;
 <a href="#citing-mlj">Citing MLJ</a> 
 </p>
 </br>
@@ -86,61 +94,6 @@ illustrated in the dependency chart below.
 *Dependency chart for MLJ repositories. Repositories with dashed
 connections do not currently exist but are planned/proposed.*
 
-### List of Wrapped Models
-
-MLJ provides access to to a wide variety of machine learning models.
-We are always looking for [help](CONTRIBUTING.md) adding new models or
-test existing ones.  Currently available models are listed below; for
-the most up-to-date list, run `using MLJ; models()`.
-
-* *experimental*: indicates the package is fairly new and/or is under active development; you can help by testing these packages and making them more robust,
-* *medium*: indicates the package is fairly mature but may benefit from optimisations and/or extra features; you can help by suggesting either,
-* *high*: indicates the package is very mature and functionalities are expected to have been fairly optimised and tested.
-
-| Package | Models | Maturity | Note
-| ------- | ------ | -------- | ----
-[Clustering.jl] | KMeans, KMedoids | high | †
-[DecisionTree.jl] | DecisionTreeClassifier, DecisionTreeRegressor, AdaBoostStumpClassifier | high | †
-[EvoTrees.jl] | EvoTreeRegressor, EvoTreeClassifier, EvoTreeCount, EvoTreeGaussian | medium | gradient boosting models
-[GLM.jl] | LinearRegressor, LinearBinaryClassifier, LinearCountRegressor | medium | †
-[LIBSVM.jl] | LinearSVC, SVC, NuSVC, NuSVR, EpsilonSVR, OneClassSVM | high | also via ScikitLearn.jl
-[LightGBM.jl] | LightGBMClassifier, LightGBMRegressor | high | 
-[MLJFlux.jl] | NeuralNetworkRegressor, NeuralNetworkClassifier, MultitargetNeuralNetworkRegressor, ImageClassifier | experimental |
[…]
