### 2020-10-12 2e42bcb
Move list of models to new section of manual; improve readme
