### 2020-10-12 9a63f6d
readme typos
