### 2020-10-12 2b95aa6
add ScientificTypes to doc/Project.toml
