### 2020-10-12 b7c9611
more doc tweaks
