### 2020-10-13 #2704 @ablaom
@henrykironde Thank you for taking the time to review this software and for the
helpful feedback.

I'm am copying my co-authors in case they wish to make additional comments.

@tlienart @vollmersj @darenasc @fkiraly

---

> **Contribution and authorship:**
>
> `What cliteria was used for authorship selection?`

Anyone making significant contribution to the design of the software.

>
> **License:**
>
> `Could you edit the licence file and remove '>'`

Done.

>
> **The paper should be between 250-1000 words:**
>
> `I will need to obain a recommendation for the paper length from the team.`

It would be good to know if there's some flexibility here. This is a
relatively large collection of packages. (You can get a sense of this
by looking at the size of the documentation, and from the [code organisation](https://github.com/alan-turing-institute/MLJ.jl/blob/master/ORGANIZATION.md)).  Of course, the paper is
not trying to duplicate documentation but is focused on outlining the
high-level design and feature set.

> **Functionality documentation:**
>
> ```
> Functions should have docs strings which are automatically turned into documentation.
> A good example is listed
> ```

Sorry, not sure I follow. Function docstrings are indeed automatically
interpolated into [the
documentation](https://alan-turing-institute.github.io/MLJ.jl/dev/)
(using Documenter.jl). The only function docstrings missing that I am
aware of are those for "private" methods. Are you suggesting adding
these to the docs?

In any case, an index of all referenced function docstrings has been
been added to a new [section](https://alan-turing-institute.github.io/MLJ.jl/dev/api/) of the docs.

>
> ** Installation instructions:**
>
> ```
> These can be made more clear.
> The Installation instructions shown in the readme file after
> the intro tend to usually clear to users.
> ```

> ** Example usage:**
>
> ```
> I expect a user to be able to run a sample example by reading the `README` file in a few minutes.
> I recommend that you move the `List of Wrapped Models` to any other location in the docs, but provide the link on the `README`
> Then focus on a sample run/s. I have tried to run the code and faced some difficulties:
> ```
>
> ** Community guidelines:**
>
> ```
> I should be able to see a developer doc section.
> In case I want to clone this repo and change a few things.
> I should be able to install the source from the current directory and there after running the code with the changes.
> I expect the user to also have guidelines on how to run tests.
> ```
>

Some context: Following a suggestion in
https://github.com/alan-turing-institute/MLJ.jl/issues/530 it was
decided that the landing page for general users of this software
should be the [GitHub Pages
site](https://alan-turing-institute.github.io/MLJ.jl/dev/), rather
than the README.md, which is common for larger packages such as this
one.

To address the issues you raise here, we've made the following changes to the
readme and manual (github pages) (see the October 12th/13th
[commits](https://github.com/alan-turing-institute/MLJ.jl/commits/master))

- The ["lightning tour"](https://alan-turing-institute.github.io/MLJ.jl/dev/#Lightning-tour-1) is now self-contained, in the sense that code required for
  installing all needed libraries is included.

-  A link to the installation instruction for Julia language has been added to that section.

-  The
   [Installation](https://alan-turing-institute.github.io/MLJ.jl/dev/#Installation-1) now provides instructions on how to test the MLJ installation.

-  As suggested, the list of supported models has been moved from the
   README.md to a [new section](https://alan-turing-institute.github.io/MLJ.jl/dev/list_of_supported_models/)  of the manual. The reamining information in the README.md is only
   for developers of the core MLJ ecosystem, as now explained
   [there](https://github.com/alan-turing-institute/MLJ.jl). Developers who specifically want to integrate new third party
   models into MLJ are now directed from the README.md to the relevant
   section of the manual (which includes a [detailed
   description](https://alan-turing-institute.github.io/MLJ.jl/dev/adding_models_for_general_use/)  of the model API). Conversely, the manual now has a link at the top of
   the page "For Developers" which directs the to the
   README/reposistory. A [Customizing
   Behavior](https://github.com/alan-turing-institute/MLJ.jl/#customizing-behavior) section has been added to the readme for those that want to replace
   a component library with a local fork they have edited.
