### 2020-10-13 #2704 @ablaom
@degleris1 Thank you for taking the time to review this software and for the
positive feedback.

I'm am copying my co-authors in case they wish to make additional comments.

@tlienart @vollmersj @darenasc @fkiraly

---

> Some really minor comments. These are things that you by no means have to address.
>
> * As has been pointed out, the installation instructions are 2 link clicks away from the README. I don't have a problem with it, but it might be nice to have something right in the README, e.g.
>

See my preceding comment about github pages being the landing page for general users. Also, the lightning tour at top of landing page now includes installation instructions.

>
> * In the [FAQ](https://alan-turing-institute.github.io/MLJ.jl/dev/frequently_asked_questions/), I noticed you used "his code" to refer to the user. It might be nice to change it to "their code" or "the user's code." As far as I can tell, it's actually the only gendered pronoun in entire docs / paper.

Good catch. Corrected thanks.

> * Is there a good reason why I have to type `verbosity` to `Int`? It seems a little bit random from a user perspective, but I'm sure there's a good reason for it.

Very observant. There used to be a method ambiguity, which has now been removed. I've updated the manual accordingly.
