### 2020-10-13 #2704 @terrytangyuan
@openjournals/joss-eics Is there any requirement on the length of the JOSS paper?
