### 2020-10-13 #2704 @danielskatz
yes, see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain

we typically are somewhat flexible on the upper bound - but this paper is too long. Detailed content should be in the documentation, not the paper.  The paper is not the user guide.
