### 2020-10-13 8e561de
update manual for new serialization API
forgotten stuff
