### 2020-10-16 #2704 @ablaom
> we typically are somewhat flexible on the upper bound - but this paper is too long. Detailed content should be in the documentation, not the paper. The paper is not the user guide.

I don't agree that the paper duplicates the documentation. Rather it discusses the high level design, focusing on those aspects that are novel or different from other machine learning toolboxes.

That said, after more carefully reading the link above, it seems this forum is not intended for design discussions. Contributions are more in the way of "announcements" than "papers". Yes?

 Over the past few days I have worked to condense the material: The arXiv version is 7.5 pages (excluding references) - the [revised version](https://github.com/ablaom/MLJDesignPaper/blob/master/arXiv2/blaom.pdf) in the same format is 5.5 pages. If this is an acceptable length - and you are happy with a "design paper" after all, let me know. Otherwise we will withdraw the manuscript and try a more suitable forum for publication.

In any case, my thanks again to the referees and JOSS team for your time.
