### 2020-10-16 #2704 @danielskatz
Right - JOSS papers are not about the process of developing the software, they are about the produced software, aimed at users of the software.
