### 2020-10-19 #2704 @henrykironde
Yes, the last concern is on `Contribution and authorship`, I wanted to get feedback on the criteria for authorship that was used.
