### 2020-10-19 #2704 @terrytangyuan
@ablaom Could you share your criteria on the authorship of this paper?
