### 2020-10-19 #2704 @ablaom
This was answered above.

> **Contribution and authorship:**

> What cliteria was used for authorship selection?
Anyone making significant contribution to the design of the software.
