### 2020-10-19 bc0b1f2
update list of models and #672
