+++ README.md
@@ -45,7 +45,9 @@ ecosystem, whose organization is described further below.
 
 MLJ (Machine Learning in Julia) is a toolbox written in Julia
 providing a common interface and meta-algorithms for selecting,
-tuning, evaluating, composing and comparing machine learning models
+tuning, evaluating, composing and comparing machine over [150 machine
+learning
+models](https://alan-turing-institute.github.io/MLJ.jl/dev/list_of_supported_models/)
 written in Julia and other languages.  MLJ is released under the MIT
 licensed and sponsored by the [Alan Turing
 Institute](https://www.turing.ac.uk/).
