### 2020-10-19 #2704 @henrykironde
Thanks @ablaom, I do think the author list needs to be improved. Comparing the these individuals
https://github.com/alan-turing-institute/MLJ.jl/graphs/contributors
`darenasc` https://github.com/alan-turing-institute/MLJ.jl/commits?author=darenasc
`OkonSamuel` https://github.com/alan-turing-institute/MLJ.jl/commits?author=OkonSamuel
`vollmersj` https://github.com/alan-turing-institute/MLJ.jl/commits?author=vollmersj

Based on the code functionality and bugs fixes (Not including documentation, and number of commits)I do think
the package author=OkonSamuel should also be among the list of authors unless the author declines. Let me know what you think
