### 2020-10-19 #2704 @ablaom
You need of course to look at [all the repos](). There are a dozen or so now. The core code lives at MLJBase.j, the largest repo. OkonSamuel has indeed made substantial contributions to the project - maintenance, issue resolution, new model interfaces, adding functionality, and a smaller amount of design work. As I said, the criterion for inclusion was around substantial design rather than code (one author contributed very substantially to design but made, I think, 0 commits).

I would be very happy to add OkonSamuel but need to check with the other authors.
