### 2020-10-20 #2704 @ablaom
Will do.

If there is no objection, I will also update the markdown to match the [shortened version](https://github.com/ablaom/MLJDesignPaper/blob/master/arXiv2/blaom.pdf) discussed above.
