### 2020-10-21 73664d2
shorten the markdown version of paper
