+++ paper/paper.md
@@ -47,34 +47,34 @@ bibliography: paper.bib
 
 # Introduction
 
-Statistical modelling, and the building of complex modelling
-pipelines, is a cornerstone of modern data science. Most experienced
-data scientists rely on high-level open source modelling toolboxes -
+Statistical modeling, and the building of complex modeling pipelines,
+is a cornerstone of modern data science. Most experienced
+data scientists rely on high-level open source modeling toolboxes -
 such as sckit-learn [@Pedregosa2001; @Buitinck2013] (Python); Weka
 [@Holmes1994] (Java); mlr [@BischlEtal2016] and caret [@Kuhn2008]
 (R) - for quick blueprinting, testing, and creation of
-deployment-ready models. They do this by providing a common interface
+deployment-ready models.  They do this by providing a common interface
 to atomic components, from an ever-growing model zoo, and by providing
 the means to incorporate these into complex work-flows. Practitioners
-are able to build increasingly sophisticated composite models, as
+are wanting to build increasingly sophisticated composite models, as
 exemplified in the strategies of top contestants in machine learning
 competitions such as Kaggle.
 
-MLJ (Machine Learning in Julia) [@MLJ] is a toolbox written in Julia
-that provides a common interface and meta-algorithms for selecting,
-tuning, evaluating, composing and comparing machine learning model 
+MLJ (Machine Learning in Julia) [@MLJ] is a toolbox written in
+Julia that provides a common interface and meta-algorithms for
+selecting, tuning, evaluating, composing and comparing machine model
 implementations written in Julia and other languages. More broadly,
-the MLJ project hopes to bring cohesion and focus to a number of emerging
-and existing, but previously disconnected, machine learning algorithms
-and tools of high quality, written in Julia. A welcome corollary of
-this activity will be increased cohesion and synergy within the
-talent-rich communities developing these tools. 
+the MLJ project hopes to bring cohesion and focus to a number of
+emerging and existing, but previously disconnected, machine learning
+algorithms and tools of high quality, written in Julia. A welcome
+corollary of this activity will be increased cohesion and synergy
+within the talent-rich communities developing these tools.
 
-In addition to
-other novelties outlined below, MLJ aims to provide first-in-class
-model composition capabilities.  Guiding goals of the MLJ project have
-been usability, interoperability, extensibility, code transparency,
-and reproducibility.
+In addition to other novelties outlined below, MLJ aims to provide
+first-in-class model composition capabilities. Guiding goals of the
+MLJ project have been usability, interoperability, extensibility, code
+transparency, and reproducibility.
+
 
 ## Why Julia?
 
@@ -93,168 +93,105 @@ accelerated through a dramatically reduced software development cycle.
 
 As an example of the productivity boost provided by the
 single-language paradigm, we cite the DifferentialEquations.jl package
-[@RackauckasNie2017], which, in a few short years of
-development by a small team of domain experts, became the best package
-in its class [@Rackauckas2017].
+[@RackauckasNie2017], which, in a few short years of development
+by a small team of domain experts, became the best package in its
+class [@Rackauckas2017].
 
-Another major advantange of a single-language solution is the ability to
-automatically differentiate (AD) functions from their code
+Another major advantage of a single-language solution is the ability
+to automatically differentiate (AD) functions from their code
 representations. The Flux.jl package [@Innes2018], for example,
 already makes use of AD to allow unparalleled flexibility in neural
-network design .
+network design.
 
 As a new language, Julia is high-performance computing-ready, and its
-superlative metaprogramming features allow developers to create
-domain-specific syntax for user interaction. 
+superlative meta-programming features allow developers to create
[…]
