+++ paper/paper.bib
@@ -1783,6 +1783,15 @@ year = "2013"
     doi = {10.5334/jors.151}
 }
 
+@misc{ScientificTypes,
+    author = {{Anthony Blaom and collaborators}},
+    title = {{ScientificTypes.jl: An API for dispatching on the "scientific" type of data instead of the machine type}},
+    year = {2019},
+    publisher = {GitHub},
+    journal = {​GitHub repository},
+    Adsurl = {​https://alan-turing-institute.github.io/ScientificTypes/}
+}
+
 @misc{Quinn,
     author = {J. Quinn},
     title = {Tables.jl: {A}n interface for tables in {J}ulia},
