### 2020-10-21 df3d074
remove redundant references from paper.bib
