+++ paper/paper.bib
@@ -1,341 +1,3 @@
-@inproceedings{beygelzimer2005weighted,
-  title={Weighted one-against-all},
-  author={Beygelzimer, Alina and Langford, John and Zadrozny, Bianca},
-  booktitle={American Association for Artificial Intelligence (AAAI)},
-  pages={720--725},
-  year={2005}
-}
-@article{burman1992data,
-  title={Data-Dependent Estimation Of Prediction Functions},
-  author={Burman, Prabir and Nolan, Deborah},
-  journal={Journal of Time Series Analysis},
-  volume={13},
-  number={3},
-  pages={189--207},
-  year={1992},
-  publisher={Wiley Online Library}
-}
-
-@book{mcquarrie1998regression,
-  title={Regression and time series model selection},
-  author={McQuarrie, Allan DR and Tsai, Chih-Ling},
-  year={1998},
-  publisher={World Scientific}
-}
-@article{bergmeir2018note,
-  title={A note on the validity of cross-validation for evaluating autoregressive time series prediction},
-  author={Bergmeir, Christoph and Hyndman, Rob J and Koo, Bonsoo},
-  journal={Computational Statistics \& Data Analysis},
-  volume={120},
-  pages={70--83},
-  year={2018},
-  publisher={Elsevier}
-}
-@article{kunst2008cross,
-  title={Cross validation of prediction models for seasonal time series by parametric bootstrapping},
-  author={Kunst, Robert M},
-  journal={Austrian Journal of Statistics},
-  volume={37},
-  number={3\&4},
-  pages={271--284},
-  year={2008}
-}
-@article{racine2000consistent,
-  title={Consistent cross-validatory model-selection for dependent data: hv-block cross-validation},
-  author={Racine, Jeff},
-  journal={Journal of econometrics},
-  volume={99},
-  number={1},
-  pages={39--61},
-  year={2000},
-  publisher={Elsevier}
-}
-@article{burman1994cross,
-  title={A cross-validatory method for dependent data},
-  author={Burman, Prabir and Chow, Edmond and Nolan, Deborah},
-  journal={Biometrika},
-  volume={81},
-  number={2},
-  pages={351--358},
-  year={1994},
-  publisher={Oxford University Press}
-}
-@book{gyorfi2013nonparametric,
-  title={Nonparametric curve estimation from time series},
-  author={Gy{\"o}rfi, L{\'a}zl{\'o} and H{\"a}rdle, Wolfgang and Sarda, Pascal and Vieu, Philippe},
-  volume={60},
-  year={2013},
-  publisher={Springer}
-}
-@incollection{beygelzimer2008machine,
-  title={Machine learning techniques—reductions between prediction quality metrics},
-  author={Beygelzimer, Alina and Langford, John and Zadrozny, Bianca},
-  booktitle={Performance Modeling and Engineering},
-  pages={3--28},
-  year={2008},
-  publisher={Springer}
-}
-@article{deng2013time,
-  title={A time series forest for classification and feature extraction},
-  author={Deng, Houtao and Runger, George and Tuv, Eugene and Vladimir, Martyanov},
-  journal={Information Sciences},
-  volume={239},
-  pages={142--153},
-  year={2013},
-  publisher={Elsevier}
-}
-@article{terpilowski2019scikit,
-  title={scikit-posthocs: Pairwise multiple comparison tests in Python},
-  author={Terpilowski, M},
-  journal={Journal of Open Source Software},
-  volume={4},
-  pages={1169},
-  year={2019}
-}
-@book{Lutkepohl2005,
-author = {L{\"{u}}tkepohl, Helmut},
-publisher = {Springer Science {\&} Business Media},
[…]
