### 2020-10-21 7373526
add DOI for papers where available
