+++ paper/paper.bib
@@ -9,6 +9,7 @@ title = {{WEKA: {A} machine learning workbench}},
 Adsurl = {http://ieeexplore.ieee.org/document/396988/},
 year = {1994}
 }
+
 @article{Gressmann2018,
     title={Probabilistic supervised learning},
     author={Frithjof Gressmann and Franz J. Király and Bilal Mateen and Harald Oberhauser},
@@ -150,7 +151,7 @@ year = {2011}
 
 @misc{MLJ,
     author = {Blaom, Anthony et al.},
-    title = {{MLJ: {A} machine learning framework for {J}ulia}},
+    title = {{MLJ: A machine learning framework for Julia}},
     year = {2019},
     publisher = {​GitHub},
     journal = {​GitHub repository},
