### 2020-10-21 5d65053
fix formatting of table output of evaluate
