+++ paper/paper.bib
@@ -6,7 +6,7 @@ isbn = {0-7803-2404-8},
 pages = {357--361},
 publisher = {IEEE},
 title = {{WEKA: {A} machine learning workbench}},
-Adsurl = {http://ieeexplore.ieee.org/document/396988/},
+url = {http://ieeexplore.ieee.org/document/396988/},
 year = {1994}
 }
 
@@ -24,7 +24,7 @@ journal = {The Journal of Machine Learning Research},
 pages = {2825--2830},
 publisher = {MIT Press},
 title = {{Scikit-learn: Machine Learning in Python}},
-Adsurl = {https://dl.acm.org/citation.cfm?id=2078195},
+url = {https://dl.acm.org/cditation.cfm?id=2078195},
 volume = {12},
 year = {2011}
 }
@@ -62,7 +62,7 @@ year = {2011}
    MRCLASS = {68N15 (65Y05 97P40)},
   MRNUMBER = {3605826},
        DOI = {10.1137/141000671},
-       Adsurl = {https://doi.org/10.1137/141000671},
+       url = {https://doi.org/10.1137/141000671},
 }
 
 @article{Blaom_I,
@@ -74,7 +74,7 @@ year = {2011}
 
 @article{Innes2018,
   doi = {10.21105/joss.00602},
-  Adsurl = {https://doi.org/10.21105/joss.00602},
+  url = {https://doi.org/10.21105/joss.00602},
   year = {2018},
   publisher = {The Open Journal},
   volume = {3},
@@ -97,7 +97,7 @@ year = {2011}
    issn = {1548-7660},
    pages = {1--26},
    doi = {10.18637/jss.v028.i05},
-   Adsurl = {https://www.jstatsoft.org/v028/i05}
+   url = {https://www.jstatsoft.org/v028/i05}
 }
 
 @software{LinEtal2020,
@@ -146,7 +146,7 @@ year = {2011}
     year = {2014},
     publisher = {​GitHub},
     journal = {​GitHub repository},
-    Adsurl = {https://github.com/JuliaData/CategoricalArrays.jl}
+    url = {https://github.com/JuliaData/CategoricalArrays.jl}
 }
 
 @misc{MLJ,
@@ -155,7 +155,7 @@ year = {2011}
     year = {2019},
     publisher = {​GitHub},
     journal = {​GitHub repository},
-    Adsurl = {https://github.com/alan-turing-institute/MLJ.jl}
+    url = {https://github.com/alan-turing-institute/MLJ.jl}
 }
 
 @misc{MLJdocs,
@@ -164,7 +164,7 @@ year = {2011}
     year = {2020},
     publisher = {​GitHub},
     journal = {​GitHub pages},
-    Adsurl = {https://alan-turing-institute.github.io/MLJ.jl/dev/}
+    url = {https://alan-turing-institute.github.io/MLJ.jl/dev/}
 }
 
 @misc{MLJTuning,
@@ -173,7 +173,7 @@ year = {2011}
     year = {2020},
     publisher = {​GitHub},
     journal = {​GitHub repository},
-    Adsurl = {https://alan-turing-institute.github.io/MLJTuning/}
+    url = {https://alan-turing-institute.github.io/MLJTuning/}
 }
 
 @misc{MLJtutorials,
@@ -182,7 +182,7 @@ year = {2011}
     year = {2020},
     publisher = {​GitHub},
     journal = {​GitHub pages},
-    Adsurl = {https://alan-turing-institute.github.io/MLJTutorials/}
+    url = {https://alan-turing-institute.github.io/MLJTutorials/}
 }
 
 @misc{Rackauckas2017,
@@ -191,7 +191,7 @@ year = {2011}
     year = {2017},
     publisher = {​C. Rackauckas},
[…]
