### 2020-10-21 9ab13ac
restore mysterious Adsurl
