+++ paper/paper.bib
@@ -6,7 +6,7 @@ isbn = {0-7803-2404-8},
 pages = {357--361},
 publisher = {IEEE},
 title = {{WEKA: {A} machine learning workbench}},
-url = {http://ieeexplore.ieee.org/document/396988/},
+Adurl = {http://ieeexplore.ieee.org/document/396988/},
 year = {1994}
 }
 
@@ -24,7 +24,7 @@ journal = {The Journal of Machine Learning Research},
 pages = {2825--2830},
 publisher = {MIT Press},
 title = {{Scikit-learn: Machine Learning in Python}},
-url = {https://dl.acm.org/cditation.cfm?id=2078195},
+Adurl = {https://dl.acm.org/cditation.cfm?id=2078195},
 volume = {12},
 year = {2011}
 }
@@ -45,7 +45,7 @@ year = {2011}
   volume  = {17},
   number  = {170},
   pages   = {1-5},
-  url     = {http://jmlr.org/papers/v17/15-066.html}
+  Adurl     = {http://jmlr.org/papers/v17/15-066.html}
 }
 
 @article {BezansonEtal2017,
@@ -62,7 +62,7 @@ year = {2011}
    MRCLASS = {68N15 (65Y05 97P40)},
   MRNUMBER = {3605826},
        DOI = {10.1137/141000671},
-       url = {https://doi.org/10.1137/141000671},
+       Adurl = {https://doi.org/10.1137/141000671},
 }
 
 @article{Blaom_I,
@@ -74,7 +74,7 @@ year = {2011}
 
 @article{Innes2018,
   doi = {10.21105/joss.00602},
-  url = {https://doi.org/10.21105/joss.00602},
+  Adurl = {https://doi.org/10.21105/joss.00602},
   year = {2018},
   publisher = {The Open Journal},
   volume = {3},
@@ -97,7 +97,7 @@ year = {2011}
    issn = {1548-7660},
    pages = {1--26},
    doi = {10.18637/jss.v028.i05},
-   url = {https://www.jstatsoft.org/v028/i05}
+   Adurl = {https://www.jstatsoft.org/v028/i05}
 }
 
 @software{LinEtal2020,
@@ -137,7 +137,7 @@ year = {2011}
   publisher    = {Zenodo},
   version      = {v0.23.2},
   doi          = {10.5281/zenodo.3730565},
-  url          = {https://doi.org/10.5281/zenodo.3730565}
+  Adurl          = {https://doi.org/10.5281/zenodo.3730565}
 }
 
 @misc{CategoricalArrays,
@@ -146,7 +146,7 @@ year = {2011}
     year = {2014},
     publisher = {​GitHub},
     journal = {​GitHub repository},
-    url = {https://github.com/JuliaData/CategoricalArrays.jl}
+    Adurl = {https://github.com/JuliaData/CategoricalArrays.jl}
 }
 
 @misc{MLJ,
@@ -155,7 +155,7 @@ year = {2011}
     year = {2019},
     publisher = {​GitHub},
     journal = {​GitHub repository},
-    url = {https://github.com/alan-turing-institute/MLJ.jl}
+    Adurl = {https://github.com/alan-turing-institute/MLJ.jl}
 }
 
 @misc{MLJdocs,
@@ -164,7 +164,7 @@ year = {2011}
     year = {2020},
     publisher = {​GitHub},
     journal = {​GitHub pages},
-    url = {https://alan-turing-institute.github.io/MLJ.jl/dev/}
+    Adurl = {https://alan-turing-institute.github.io/MLJ.jl/dev/}
 }
 
 @misc{MLJTuning,
@@ -173,7 +173,7 @@ year = {2011}
     year = {2020},
     publisher = {​GitHub},
     journal = {​GitHub repository},
[…]
