+++ paper/paper.bib
@@ -39,7 +39,7 @@ year = {2011}
 
 @article{BischlEtal2016,
   author  = {Bernd Bischl and Michel Lang and Lars Kotthoff and Julia Schiffner and Jakob Richter and Erich Studerus and Giuseppe Casalicchio and Zachary M. Jones},
-  title   = {mlr: Machine Learning in {R}},
+  title   = {{mlr: Machine Learning in R}},
   journal = {Journal of Machine Learning Research},
   year    = {2016},
   volume  = {17},
