### 2020-10-21 #2704 @ablaom
@terrytangyuan I have:

- removed the plethora of redundant bib entries in paper.bib
- added doi entries where these exist
- updated the markdown paper.md to the shorter version

Note that url hypertext links do not appear in the bibliography. We have replaced the `url=...` entries in paper.bib with `adsurl=...` entries because that is what we see in the JOSS template (but I have no idea what "adsurl" means) and because I get compile errors if I use `url`. Are you able  to shed light on this?
