### 2020-10-21 #2704 @terrytangyuan
I don't have much knowledge in the JOSS template. cc'ing @openjournals/joss-eics to help answer the question above on `adsurl`.
