### 2020-10-22 #2704 @arfon
`url` should be used `adsurl` shouldn't be necessary (this is an astronomy-specific service that the author of the example paper used).

That said, there's something really weird going on with your BibTeX file here @ablaom. When I try and compile your locally I also get weird errors, but if I reduce the BibTeX file to _just_ this entry:

```
@misc{Quinn,
  author = {J. Quinn},
  title = {Tables.jl: {A}n interface for tables in {J}ulia},
  year = {2020},
  publisher = {​GitHub},
  journal = {​GitHub repository},
  url = {https://github.com/JuliaData/Tables.jl}
}
```

...then this entry works fine. In getting this to work I had to remove all of the whitespace and reformat the entries so I'm wondering if there's some weird hidden characters in your BibTeX file somewhere.

I don't have time to dig into this further at this point sorry but can try and take another look over the weekend.

Also, to help you debug locally, this GitHub Action we've been working on may help you: https://github.com/marketplace/actions/open-journals-pdf-generator
