+++ paper/paper.bib
@@ -1,40 +1,40 @@
 @inproceedings{Holmes1994,
-author = {Holmes, G. and Donkin, A. and Witten, I.H.},
-booktitle = {Proceedings of ANZIIS '94 - Australian New Zealnd Intelligent Information Systems Conference},
-doi = {10.1109/ANZIIS.1994.396988},
-isbn = {0-7803-2404-8},
-pages = {357--361},
-publisher = {IEEE},
-title = {{WEKA: {A} machine learning workbench}},
-Adurl = {http://ieeexplore.ieee.org/document/396988/},
-year = {1994}
+  author    = {Holmes, G. and Donkin, A. and Witten, I.H.},
+  booktitle = {Proceedings of ANZIIS '94 - Australian New Zealnd Intelligent Information Systems Conference},
+  doi       = {10.1109/ANZIIS.1994.396988},
+  isbn      = {0-7803-2404-8},
+  pages     = {357--361},
+  publisher = {IEEE},
+  title     = {{WEKA: {A} machine learning workbench}},
+  adurl     = {http://ieeexplore.ieee.org/document/396988/},
+  year      = {1994}
 }
 
 @article{Gressmann2018,
-    title={Probabilistic supervised learning},
-    author={Frithjof Gressmann and Franz J. Király and Bilal Mateen and Harald Oberhauser},
-    year={2018},
-    journal= {ArXiv},
-    volume={1801.00753}
+  title   = {Probabilistic supervised learning},
+  author  = {Frithjof Gressmann and Franz J. Király and Bilal Mateen and Harald Oberhauser},
+  year    = {2018},
+  journal = {ArXiv},
+  volume  = {1801.00753}
 }
 @article{Pedregosa2001,
-author = {Pedregosa, Fabian and Varoquaux, Ga{\"{e}}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, {\'{E}}douard},
-file = {:Users/mloning/Library/Application Support/Mendeley Desktop/Downloaded/Pedregosa et al. - 2001 - Scikit-learn Machine Learning in Python.pdf:pdf},
-journal = {The Journal of Machine Learning Research},
-pages = {2825--2830},
-publisher = {MIT Press},
-title = {{Scikit-learn: Machine Learning in Python}},
-Adurl = {https://dl.acm.org/cditation.cfm?id=2078195},
-volume = {12},
-year = {2011}
+  author    = {Pedregosa, Fabian and Varoquaux, Ga{\"{e}}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, {\'{E}}douard},
+  file      = {:Users/mloning/Library/Application Support/Mendeley Desktop/Downloaded/Pedregosa et al. - 2001 - Scikit-learn Machine Learning in Python.pdf:pdf},
+  journal   = {The Journal of Machine Learning Research},
+  pages     = {2825--2830},
+  publisher = {MIT Press},
+  title     = {{Scikit-learn: Machine Learning in Python}},
+  adurl     = {https://dl.acm.org/cditation.cfm?id=2078195},
+  volume    = {12},
+  year      = {2011}
 }
 
[…]
