### 2020-10-28 4252a21
verification of correct urls in paper.bib
