+++ paper/paper.bib
@@ -24,7 +24,7 @@
   pages     = {2825--2830},
   publisher = {MIT Press},
   title     = {{Scikit-learn: Machine Learning in Python}},
-  url       = {https://dl.acm.org/cditation.cfm?id=2078195},
+  url       = {https://dl.acm.org/doi/10.5555/1953048.2078195},
   volume    = {12},
   year      = {2011}
 }
@@ -131,14 +131,15 @@
                   Richard Reeve and
                   Kai Xu and
                   David Widmann},
-  title     = {Distributions.jl: {A} {J}ulia package for probability distributions and associated functions},
+  title     = {JuliaStats/Distributions.jl: {A} {J}ulia package for probability distributions and associated functions},
   month     = mar,
   year      = 2020,
   publisher = {Zenodo},
   version   = {v0.23.2},
   doi       = {10.5281/zenodo.3730565},
-  url       = {https://doi.org/10.5281/zenodo.3730565}
+  url       = {http://doi.org/10.5281/zenodo.3730565}
 }
+
 
 @misc{CategoricalArrays,
   author    = {Bouchet-Valat, Milan et al.},
@@ -173,16 +174,16 @@
   year      = {2020},
   publisher = {GitHub},
   journal   = {GitHub repository},
-  url       = {https://alan-turing-institute.github.io/MLJTuning/}
+  url       = {https://github.com/alan-turing-institute/MLJTuning.jl}
 }
 
 @misc{MLJtutorials,
   author    = {Lienart, Thibaut and Blaom, Anthony and collaborators},
-  title     = {Data Analysis and Machine Learning in {J}ulia},
+  title     = {Data Science Tutorials in {J}ulia},
   year      = {2020},
   publisher = {GitHub},
   journal   = {GitHub pages},
-  url       = {https://alan-turing-institute.github.io/MLJTutorials/}
+  url       = {https://alan-turing-institute.github.io/DataScienceTutorials.jl/}
 }
 
 @misc{Rackauckas2017,
@@ -202,7 +203,8 @@
   volume  = 5,
   number  = 1,
   pages   = {15},
-  doi     = {10.5334/jors.151}
+  doi     = {10.5334/jors.151},
+  url     = {https://openresearchsoftware.metajnl.com/articles/10.5334/jors.151/}
 }
 
 @misc{ScientificTypes,
@@ -211,7 +213,7 @@
   year      = {2019},
   publisher = {GitHub},
   journal   = {GitHub repository},
-  url       = {https://alan-turing-institute.github.io/ScientificTypes/}
+  url       = {https://github.com/alan-turing-institute/ScientificTypes.jl}
 }
 
 @misc{Quinn,
