### 2020-10-28 #2704 @ablaom
@terrytangyuan @arfon

Thanks to @darenasc I believe the bib issue is now sorted.
