### 2020-10-28 4e0b042
suspend EnsembleModel testing; MLJNearestNeighborsInterface issue 1
