+++ test/runtests.jl
@@ -12,9 +12,10 @@ end
     @test include("version.jl")
 end
 
-@testset "ensembles" begin
-    @test include("ensembles.jl")
-end
+# TODO: restore after MLJNearestNeighbors #1 is resolved:
+# @testset "ensembles" begin
+#    @test include("ensembles.jl")
+# end
 
 @testset "matching models to data" begin
     @test include("model_matching.jl")
