+++ paper/paper.bib
@@ -186,13 +186,14 @@
   url       = {https://alan-turing-institute.github.io/DataScienceTutorials.jl/}
 }
 
-@misc{Rackauckas2017,
-  author    = {C. Rackauckas},
-  title     = {A Comparison Between Differential Equation Solver Suites In {MATLAB}, {R}, {J}ulia, {P}ython, {C}, {M}athematica, {M}aple, and {F}ortran},
-  year      = {2017},
-  publisher = {C. Rackauckas},
-  journal   = {Stochastic Lifestyle (blog)},
-  url       = {https://www.stochasticlifestyle.com/comparison-differential-equation-solver-suites-matlab-r-julia-python-c-fortran/}
+@article{Rackauckas2017,
+  author  = {Christopher Rackauckas},
+  journal = {The Winnower},
+  title   = {A Comparison Between Differential Equation Solver Suites In MATLAB, R, Julia, Python, C, Mathematica, Maple, and Fortran},
+  year    = {2018},
+  month   = {08},
+  url     = {https://dx.doi.org/10.15200/winn.153459.98975},
+  doi     = {10.15200/winn.153459.98975}
 }
 
 @article{RackauckasNie2017,
