### 2020-11-03 #2704 @ablaom
@Kevin-Mattheus-Moerman

I have added the country to "University of Auckland" and "London" to "The Alan Institute". I think adding city or country to any other entry is redundant - even silly.  Eg, University College London" is obviously in London, and an unqualified "London" obviously refers to "London England" not "London, Kentucky".

Only the first two institutions on the list  are not very well-known. The New Zealand eScience institute is a national institute and not located in one NZ city.

If you literally need this information explicit, let me know.
