### 2020-11-04 #2704 @fkiraly
>and an unqualified "London" obviously refers to "London England" not "London, Kentucky".

Wait, this is *not* London, Kentucky where I'm living?
That would explain all the Grenadier Guards in the city, and the lack of banjos.
