### 2020-11-06 5cabbe9
Added LatinHypercube to tuning_models.md
