### 2020-11-06 81affc6
Added latin_hypercube docs
